src/requirements.txt
gunicorn==23.0.0
firebase_admin==6.6.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-WTF==1.2.2
google-api-core==2.24.1
google-api-python-client==2.161.0
google-cloud-logging==3.11.4
google-cloud-storage==3.0.0
grpcio==1.70.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.1.2
oauth2client==4.1.3
pyre-extensions==0.0.32
pytz==2025.1
requests==2.32.3
typing-extensions==4.12.2
Werkzeug==3.1.3
appengine-python-standard==1.1.9